Fix not shared dropbox activity + refactor activities

+++ app/src/main/java/com/infomaniak/drive/data/cache/FileController.kt
@@ -720,7 +720,7 @@ object FileController {
         return files?.let { ArrayList(it) } ?: arrayListOf()
     }
 
-    fun getFolderActivities(folder: File, page: Int, userDrive: UserDrive? = null): Map<out Int, File.LocalFileActivity> {
+    fun getFolderActivities(folder: File, page: Int, userDrive: UserDrive? = null): Map<out Int, FileActivity> {
         return getRealmInstance(userDrive).use { realm ->
             getFolderActivitiesRec(realm, folder, page, userDrive)
         }
@@ -731,11 +731,11 @@ object FileController {
         folder: File,
         page: Int,
         userDrive: UserDrive? = null
-    ): Map<out Int, File.LocalFileActivity> {
+    ): Map<out Int, FileActivity> {
         val okHttpClient = runBlocking {
             userDrive?.userId?.let { KDriveHttpClient.getHttpClient(it) } ?: HttpClient.okHttpClient
         }
-        val returnResponse = arrayMapOf<Int, File.LocalFileActivity>()
+        val returnResponse = arrayMapOf<Int, FileActivity>()
         val apiResponse = ApiRepository.getFileActivities(okHttpClient, folder, page)
         if (!apiResponse.isSuccess()) return returnResponse
 
@@ -769,33 +769,30 @@ object FileController {
         }
     }
 
-    private fun FileActivity.applyFileActivity(
-        realm: Realm,
-        returnResponse: ArrayMap<Int, File.LocalFileActivity>,
-        currentFolder: File
-    ) {
-        val fileId = this.fileId
+    private fun FileActivity.applyFileActivity(realm: Realm, returnResponse: ArrayMap<Int, FileActivity>, currentFolder: File) {
         when (this.getAction()) {
             FileActivity.FileActivityType.FILE_DELETE,
             FileActivity.FileActivityType.FILE_MOVE_OUT,
-            FileActivity.FileActivityType.FILE_TRASH -> if (returnResponse[this.fileId] == null) {
-                getParentFile(fileId = fileId, realm = realm)?.let { parent ->
-                    if (parent.id == currentFolder.id) {
-                        removeFile(fileId, customRealm = realm, recursive = false)
+            FileActivity.FileActivityType.FILE_TRASH -> {
+                if (returnResponse[fileId] == null || returnResponse[fileId]?.createdAt?.time == this.createdAt.time) { // Api fix
+                    getParentFile(fileId = fileId, realm = realm)?.let { parent ->
+                        if (parent.id == currentFolder.id) {
+                            removeFile(fileId, customRealm = realm, recursive = false)
+                        }
                     }
+                    returnResponse[fileId] = this
                 }
-                returnResponse[this.fileId] = File.LocalFileActivity.IS_DELETE
             }
             FileActivity.FileActivityType.FILE_CREATE,
             FileActivity.FileActivityType.FILE_MOVE_IN,
-            FileActivity.FileActivityType.FILE_RESTORE -> if (returnResponse[this.fileId] == null && this.file != null) {
+            FileActivity.FileActivityType.FILE_RESTORE -> if (returnResponse[fileId] == null && this.file != null) {
                 realm.where(File::class.java).equalTo(File::id.name, currentFolder.id).findFirst()?.let { realmFolder ->
                     if (!realmFolder.children.contains(this.file)) {
                         addChild(realm, realmFolder, this.file!!)
-                        returnResponse[this.fileId] = File.LocalFileActivity.IS_NEW
                     } else {
-                        updateFileFromActivity(realm, returnResponse, this, realmFolder.id)
+                        updateFileFromActivity(realm, this, realmFolder.id)
                     }
+                    returnResponse[fileId] = this
                 }
             }
             FileActivity.FileActivityType.COLLABORATIVE_FOLDER_CREATE,
@@ -809,13 +806,13 @@ object FileController {
             FileActivity.FileActivityType.FILE_SHARE_CREATE,
             FileActivity.FileActivityType.FILE_SHARE_DELETE,
             FileActivity.FileActivityType.FILE_SHARE_UPDATE,
-            FileActivity.FileActivityType.FILE_UPDATE -> if (returnResponse[this.fileId] == null) {
+            FileActivity.FileActivityType.FILE_UPDATE -> if (returnResponse[fileId] == null) {
                 if (this.file == null) {
                     removeFile(fileId, customRealm = realm, recursive = false)
-                    returnResponse[this.fileId] = File.LocalFileActivity.IS_DELETE
                 } else {
-                    updateFileFromActivity(realm, returnResponse, this, currentFolder.id)
+                    updateFileFromActivity(realm, this, currentFolder.id)
                 }
+                returnResponse[fileId] = this
             }
             else -> Unit
         }
@@ -852,18 +849,10 @@ object FileController {
         return customRealm?.let(block) ?: getRealmInstance(userDrive).use(block)
     }
 
-    private fun updateFileFromActivity(
-        realm: Realm,
-        returnResponse: ArrayMap<Int, File.LocalFileActivity>,
-        fileActivity: FileActivity,
-        folderId: Int
-    ) {
-        returnResponse[fileActivity.fileId] = File.LocalFileActivity.IS_UPDATE
-
+    private fun updateFileFromActivity(realm: Realm, fileActivity: FileActivity, folderId: Int) {
         getFileProxyById(fileActivity.fileId, customRealm = realm)?.let { file ->
             insertOrUpdateFile(realm, fileActivity.file!!, file)
         } ?: also {
-            returnResponse[fileActivity.fileId] = File.LocalFileActivity.IS_NEW
             realm.executeTransaction {
                 realm.where(File::class.java).equalTo(File::id.name, folderId).findFirst()?.children?.add(fileActivity.file)
             }
